fix many crashes

+++ app/src/main/java/com/example/xmaster/market/CoinListItemViewModel.kt
@@ -9,7 +9,7 @@ class CoinListItemViewModel(item: Coin) {
     val symbol: String
     val cmc_rank: Int
     val price: Float
-    val circulating_supply: Long
+    val circulating_supply: Float
     val percent_change_24h: Float
 
     init {
